+++ tests/apps/courses/test_cms_plugins_course.py
@@ -124,7 +124,7 @@ class CoursePluginTestCase(TestCase):
                 (
                     r'<div class="course-glimpse__organization">'
                     r'<svg role="img".*'
-                    r'<use href="#icon-pin" />'
+                    r'<use href="#icon-org" />'
                 ),
                 str(response.content),
             )
@@ -183,7 +183,7 @@ class CoursePluginTestCase(TestCase):
                 (
                     r'<div class="course-glimpse__organization">'
                     r'<svg role="img".*'
-                    r'<use href="#icon-pin" />'
+                    r'<use href="#icon-org" />'
                 ),
                 str(response.content),
             )